requirements.txt
# ── Core ─────────────────────────────────────────────────
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# ── API Server ───────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12        # required for file uploads

# ── Orchestration ────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# ── Document Parsing ────────────────────────────────────
PyMuPDF>=1.24.0
python-docx>=1.1.0

# ── Embeddings & Vector Store ────────────────────────────
sentence-transformers>=3.0.0
chromadb>=0.5.0

# ── Storage ──────────────────────────────────────────────
pymongo>=4.8.0
boto3>=1.35.0

# ── Utilities ────────────────────────────────────────────
python-dotenv>=1.0.0
dateparser>=1.2.0

# ── Testing ──────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
